refactor(benchmark): update session load benchmark
